workaround for issue / accuracy cap in 2F1()
There seems to be a bug(?) in both 2F1 and 2F1general2 that is causing loss of full Float64 accuracy for certain high inputs unless the inputs are of BigFloat type. Presumably it's happening because of some intermediate allocations in HypergeometricFunctions.jl. Either way, this provides a workaround for that so that this method remains accurate to machine precision in high ranges of a, b and c.